+++ scout/adapter/mongo/event.py
@@ -15,6 +15,51 @@ LOG = logging.getLogger(__name__)
 class EventHandler(CaseEventHandler, VariantEventHandler):
     """Class to handle events for the mongo adapter"""
 
+    def user_timeline(self, user_email, limit=100):
+        """Retrieve the last events created by a user, grouped by institute, case_name, category, verb and date
+           and ordered by descending date (from the newest). Return 100 groups of these events
+
+        Args:
+            user_email(string): email of a logged user
+            limit(int): 100 to display latest events or -1 to display all events
+
+        Returns:
+            pymongo.Cursor: query results
+        """
+        # Build the query pipeline
+        match_query = {
+            "$match": {
+                "user_id": user_email,
+            }
+        }  # Return only events created by given user
+        add_fields = {
+            "$addFields": {
+                "yearMonthDay": {"$dateToString": {"format": "%Y-%m-%d", "date": "$updated_at"}},
+            }
+        }  # Create an additional field: yearMonthDay to group and display events better later
+        group = {
+            "$group": {
+                "_id": {
+                    "institute": "$institute",
+                    "case_id": "$case",
+                    "category": "$category",
+                    "verb": "$verb",
+                    "yearMonthDay": "$yearMonthDay",
+                },
+                "count": {"$sum": 1},
+            }
+        }  # Group events by institute, case_name, category, verb and date
+        sort = {
+            "$sort": {"_id.yearMonthDay": -1, "_id.institute": 1, "_id.case_id": 1}
+        }  # Sort by date desc, institute and case asc
+
+        pipeline = [match_query, add_fields, group, sort]
+
+        if limit > 0:  # Limit number of events to be displayed
+            pipeline.append({"$limit": limit})  # 100 events by default
+
+        return self.event_collection.aggregate(pipeline)
+
     def delete_event(self, event_id):
         """Delete a event