+++ CHANGELOG.md
@@ -6,6 +6,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ##[]
 ### Added
+- Display last 50 or 500 events for a user in a timeline
 - Show dismiss count from other cases on matching variantS
 ### Changed
 - Prior ACMG classifications view is no longer limited by pathogenicity